+++ experiment_types/SpontaneousActivity.py
@@ -15,11 +15,15 @@ class SpontaneousActivity:
         save_settings_config_filename,
         exp_config_filename,
         debug,
+        preview=False,
+        save_outputs=None,
     ):
         self.experiment_id = experiment_id
         self.mouse_id = mouse_id
         self.daq = daq
         self.debug = debug
+        self.preview = bool(preview)
+        self.save_outputs = (not self.preview) if save_outputs is None else bool(save_outputs)
 
         self.status_callback = None
         self.trial_callback = None
@@ -38,7 +42,8 @@ class SpontaneousActivity:
         self.data_log_dir = None
         self.experiment_log_filename = None
         self.ni_log_filename = None
-        self._create_save_directories()
+        if self.save_outputs:
+            self._create_save_directories()
 
         self.experiment_settings_filenames = [
             self.monitor_config_filename,
@@ -51,6 +56,8 @@ class SpontaneousActivity:
             self.mouse_id,
             self.data_log_dir,
             self.experiment_settings_filenames,
+            preview=self.preview,
+            save_outputs=self.save_outputs,
         )
         self.daq.ni_log_filename = self.ni_log_filename
         self.exp_log.log['daq_sampling_rate'] = self.daq.sampling_rate
@@ -113,6 +120,9 @@ class SpontaneousActivity:
     def start_data_acquisition(self):
         if self.daq is None:
             raise Exception('Please set the daq object, it has not been set.')
+
+        if self.preview:
+            return
 
         if os.sys.platform == 'win32' and not self.debug:
             self.acquisition_running = True